pyproject.toml
[project]
name = "sflt"
version = "0.1.0"
description = "Static React site deployed with AWS CDK to CloudFront + S3"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.4.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "playwright>=1.53.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-playwright>=0.7.0",
    "ruff>=0.12.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cdk"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests", "e2e/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=cdk --cov-report=html --cov-report=term"
markers = [
    "unit: Unit tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "performance: Performance tests",
]
